requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
PyYAML>=6.0
torch>=2.1